Extract type generator to module, separate & fix tests
