+++ codegen/src/types/mod.rs
@@ -40,6 +40,9 @@ use std::collections::{
     HashSet,
 };
 
+#[cfg(test)]
+mod tests;
+
 #[derive(Debug)]
 pub struct TypeGenerator<'a> {
     root_mod_ident: Ident,
@@ -304,7 +307,7 @@ impl<'a> quote::ToTokens for ModuleType<'a> {
                                     | TypeDefPrimitive::U128
                             )
                         ) {
-                            Some(quote!( #[derive(::codec::CompactAs)] ))
+                            Some(quote!( #[derive(::subxt::codec::CompactAs)] ))
                         } else {
                             None
                         }
@@ -729,781 +732,5 @@ impl TypePathSubstitute {
             let params = &self.params;
             syn::parse_quote! ( #substitute_path< #( #params ),* > )
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use pretty_assertions::assert_eq;
-    use scale_info::{
-        meta_type,
-        Registry,
-        TypeInfo,
-    };
-
-    const MOD_PATH: &'static [&'static str] = &["subxt_codegen", "types", "tests"];
-
-    fn get_mod<'a>(
-        module: &'a Module,
-        path_segs: &[&'static str],
-    ) -> Option<&'a Module<'a>> {
-        let (mod_name, rest) = path_segs.split_first()?;
-        let mod_ident = Ident::new(mod_name, Span::call_site());
-        let module = module.children.get(&mod_ident)?;
-        if rest.is_empty() {
-            Some(module)
-        } else {
-            get_mod(module, rest)
-        }
-    }
-
-    #[test]
-    fn generate_struct_with_primitives() {
-        #[allow(unused)]
-        #[derive(TypeInfo)]
-        struct S {
-            a: bool,
-            b: u32,
-            c: char,
-        }
-
-        let mut registry = Registry::new();
-        registry.register_type(&meta_type::<S>());
-        let portable_types: PortableRegistry = registry.into();
-
-        let type_gen = TypeGenerator::new(
-            &portable_types,
-            "root",
-            Default::default(),
-            Default::default(),
-        );
-        let types = type_gen.generate_types_mod();
-        let tests_mod = get_mod(&types, MOD_PATH).unwrap();
-
-        assert_eq!(
-            tests_mod.into_token_stream().to_string(),
-            quote! {
-                pub mod tests {
-                    use super::root;
-
-                    #[derive(::codec::Encode, ::codec::Decode)]
-                    pub struct S {
-                        pub a: bool,
-                        pub b: u32,
-                        pub c: char,
-                    }
-                }
-            }
-            .to_string()
-        )
-    }
-
-    #[test]
-    fn generate_struct_with_a_struct_field() {
-        #[allow(unused)]
-        #[derive(TypeInfo)]
-        struct Parent {
-            a: bool,
-            b: Child,
-        }
-
-        #[allow(unused)]
-        #[derive(TypeInfo)]
-        struct Child {
-            a: i32,
-        }
-
-        let mut registry = Registry::new();
-        registry.register_type(&meta_type::<Parent>());
-        let portable_types: PortableRegistry = registry.into();
-
-        let type_gen = TypeGenerator::new(
-            &portable_types,
-            "root",
-            Default::default(),
-            Default::default(),
-        );
-        let types = type_gen.generate_types_mod();
-        let tests_mod = get_mod(&types, MOD_PATH).unwrap();
-
-        assert_eq!(
-            tests_mod.into_token_stream().to_string(),
-            quote! {
-                pub mod tests {
-                    use super::root;
-
-                    #[derive(::codec::Encode, ::codec::Decode)]
-                    pub struct Child {
-                        pub a: i32,
-                    }
-
-                    #[derive(::codec::Encode, ::codec::Decode)]
-                    pub struct Parent {
-                        pub a: bool,
-                        pub b: root::subxt_codegen::types::tests::Child,
-                    }
-                }
-            }
-            .to_string()
-        )
-    }
-
-    #[test]
-    fn generate_tuple_struct() {
-        #[allow(unused)]
-        #[derive(TypeInfo)]
-        struct Parent(bool, Child);
-
-        #[allow(unused)]
-        #[derive(TypeInfo)]
-        struct Child(i32);
-
-        let mut registry = Registry::new();
-        registry.register_type(&meta_type::<Parent>());
-        let portable_types: PortableRegistry = registry.into();
-
-        let type_gen = TypeGenerator::new(
-            &portable_types,
-            "root",
-            Default::default(),
-            Default::default(),
-        );
-        let types = type_gen.generate_types_mod();
-        let tests_mod = get_mod(&types, MOD_PATH).unwrap();
-
-        assert_eq!(
-            tests_mod.into_token_stream().to_string(),
-            quote! {
-                pub mod tests {
-                    use super::root;
-
-                    #[derive(::codec::Encode, ::codec::Decode)]
-                    pub struct Child(pub i32,);
-
-                    #[derive(::codec::Encode, ::codec::Decode)]
-                    pub struct Parent(pub bool, pub root::subxt_codegen::types::tests::Child,);
-                }
-            }
-            .to_string()
-        )
-    }
-
-    #[test]
-    fn derive_compact_as_for_uint_wrapper_structs() {
-        #[allow(unused)]
-        #[derive(TypeInfo)]
-        struct Su8 {
-            a: u8,
-        }
-        #[allow(unused)]
-        #[derive(TypeInfo)]
-        struct TSu8(u8);
-        #[allow(unused)]
-        #[derive(TypeInfo)]
-        struct Su16 {
-            a: u16,
-        }
-        #[allow(unused)]
-        #[derive(TypeInfo)]
-        struct TSu16(u16);
-        #[allow(unused)]
-        #[derive(TypeInfo)]
-        struct Su32 {
-            a: u32,
-        }
-        #[allow(unused)]
-        #[derive(TypeInfo)]
-        struct TSu32(u32);
-        #[allow(unused)]
-        #[derive(TypeInfo)]
-        struct Su64 {
-            a: u64,
-        }
-        #[allow(unused)]
-        #[derive(TypeInfo)]
-        struct TSu64(u64);
-        #[allow(unused)]
-        #[derive(TypeInfo)]
-        struct Su128 {
-            a: u128,
-        }
-        #[allow(unused)]
-        #[derive(TypeInfo)]
-        struct TSu128(u128);
-
-        let mut registry = Registry::new();
-        registry.register_type(&meta_type::<Su8>());
-        registry.register_type(&meta_type::<TSu8>());
-        registry.register_type(&meta_type::<Su16>());
-        registry.register_type(&meta_type::<TSu16>());
-        registry.register_type(&meta_type::<Su32>());
-        registry.register_type(&meta_type::<TSu32>());
-        registry.register_type(&meta_type::<Su64>());
-        registry.register_type(&meta_type::<TSu64>());
-        registry.register_type(&meta_type::<Su128>());
-        registry.register_type(&meta_type::<TSu128>());
-        let portable_types: PortableRegistry = registry.into();
-
-        let type_gen = TypeGenerator::new(
-            &portable_types,
-            "root",
-            Default::default(),
-            Default::default(),
-        );
-        let types = type_gen.generate_types_mod();
-        let tests_mod = get_mod(&types, MOD_PATH).unwrap();
-
-        assert_eq!(
-            tests_mod.into_token_stream().to_string(),
-            quote! {
-                pub mod tests {
-                    use super::root;
-
-                    #[derive(Debug, Eq, PartialEq, ::codec::Encode, ::codec::Decode, ::codec::CompactAs)]
-                    pub struct Su128 { pub a: u128, }
-
-                    #[derive(Debug, Eq, PartialEq, ::codec::Encode, ::codec::Decode, ::codec::CompactAs)]
-                    pub struct Su16 { pub a: u16, }
-
-                    #[derive(Debug, Eq, PartialEq, ::codec::Encode, ::codec::Decode, ::codec::CompactAs)]
-                    pub struct Su32 { pub a: u32, }
-
-                    #[derive(Debug, Eq, PartialEq, ::codec::Encode, ::codec::Decode, ::codec::CompactAs)]
-                    pub struct Su64 { pub a: u64, }
-
-                    #[derive(Debug, Eq, PartialEq, ::codec::Encode, ::codec::Decode, ::codec::CompactAs)]
-                    pub struct Su8 { pub a: u8, }
-
-                     #[derive(Debug, Eq, PartialEq, ::codec::Encode, ::codec::Decode, ::codec::CompactAs)]
-                    pub struct TSu128(pub u128,);
-
-                    #[derive(Debug, Eq, PartialEq, ::codec::Encode, ::codec::Decode, ::codec::CompactAs)]
-                    pub struct TSu16(pub u16,);
-
-                    #[derive(Debug, Eq, PartialEq, ::codec::Encode, ::codec::Decode, ::codec::CompactAs)]
-                    pub struct TSu32(pub u32,);
-
-                    #[derive(Debug, Eq, PartialEq, ::codec::Encode, ::codec::Decode, ::codec::CompactAs)]
-                    pub struct TSu64(pub u64,);
-
-                    #[derive(Debug, Eq, PartialEq, ::codec::Encode, ::codec::Decode, ::codec::CompactAs)]
-                    pub struct TSu8(pub u8,);
-                }
-            }
-                .to_string()
-        )
-    }
-
-    #[test]
-    fn generate_enum() {
-        #[allow(unused)]
-        #[derive(TypeInfo)]
-        enum E {
-            A,
-            B(bool),
-            C { a: u32 },
-        }
-
-        let mut registry = Registry::new();
-        registry.register_type(&meta_type::<E>());
-        let portable_types: PortableRegistry = registry.into();
-
-        let type_gen = TypeGenerator::new(
-            &portable_types,
-            "root",
-            Default::default(),
-            Default::default(),
-        );
-        let types = type_gen.generate_types_mod();
-        let tests_mod = get_mod(&types, MOD_PATH).unwrap();
-
-        assert_eq!(
-            tests_mod.into_token_stream().to_string(),
-            quote! {
-                pub mod tests {
-                    use super::root;
-                    #[derive(::codec::Encode, ::codec::Decode)]
-                    pub enum E {
-                        A,
-                        B (bool,),
-                        C { a: u32, },
-                    }
-                }
-            }
-            .to_string()
-        )
-    }
-
-    #[test]
-    fn generate_array_field() {
-        #[allow(unused)]
-        #[derive(TypeInfo)]
-        struct S {
-            a: [u8; 32],
-        }
-
-        let mut registry = Registry::new();
-        registry.register_type(&meta_type::<S>());
-        let portable_types: PortableRegistry = registry.into();
-
-        let type_gen = TypeGenerator::new(
-            &portable_types,
-            "root",
-            Default::default(),
-            Default::default(),
-        );
-        let types = type_gen.generate_types_mod();
-        let tests_mod = get_mod(&types, MOD_PATH).unwrap();
-
-        assert_eq!(
-            tests_mod.into_token_stream().to_string(),
-            quote! {
-                pub mod tests {
-                    use super::root;
-                    #[derive(::codec::Encode, ::codec::Decode)]
-                    pub struct S {
-                        pub a: [u8; 32usize],
-                    }
-                }
-            }
-            .to_string()
-        )
-    }
-
-    #[test]
-    fn option_fields() {
-        #[allow(unused)]
-        #[derive(TypeInfo)]
-        struct S {
-            a: Option<bool>,
-            b: Option<u32>,
-        }
-
-        let mut registry = Registry::new();
-        registry.register_type(&meta_type::<S>());
-        let portable_types: PortableRegistry = registry.into();
-
-        let type_gen = TypeGenerator::new(
-            &portable_types,
-            "root",
-            Default::default(),
-            Default::default(),
-        );
-        let types = type_gen.generate_types_mod();
-        let tests_mod = get_mod(&types, MOD_PATH).unwrap();
-
-        assert_eq!(
-            tests_mod.into_token_stream().to_string(),
-            quote! {
-                pub mod tests {
-                    use super::root;
-                    #[derive(::codec::Encode, ::codec::Decode)]
-                    pub struct S {
-                        pub a: Option<bool>,
-                        pub b: Option<u32>,
-                    }
-                }
-            }
-            .to_string()
-        )
-    }
-
-    #[test]
-    fn box_fields_struct() {
-        // todo: [AJ] remove hack for Box and make no_std compatible using `alloc::Box`
-
-        use std::boxed::Box;
-
-        #[allow(unused)]
-        #[derive(TypeInfo)]
-        struct S {
-            a: std::boxed::Box<bool>,
-            b: Box<u32>,
-        }
-
-        let mut registry = Registry::new();
-        registry.register_type(&meta_type::<S>());
-        let portable_types: PortableRegistry = registry.into();
-
-        let type_gen = TypeGenerator::new(
-            &portable_types,
-            "root",
-            Default::default(),
-            Default::default(),
-        );
-        let types = type_gen.generate_types_mod();
-        let tests_mod = get_mod(&types, MOD_PATH).unwrap();
-
-        assert_eq!(
-            tests_mod.into_token_stream().to_string(),
-            quote! {
-                pub mod tests {
-                    use super::root;
-                    #[derive(::codec::Encode, ::codec::Decode)]
-                    pub struct S {
-                        pub a: std::boxed::Box<bool>,
-                        pub b: std::boxed::Box<u32>,
-                    }
-                }
-            }
-            .to_string()
-        )
-    }
-
-    #[test]
-    fn box_fields_enum() {
-        use std::boxed::Box;
-
-        #[allow(unused)]
-        #[derive(TypeInfo)]
-        enum E {
-            A(Box<bool>),
-            B { a: Box<u32> },
-        }
-
-        let mut registry = Registry::new();
-        registry.register_type(&meta_type::<E>());
-        let portable_types: PortableRegistry = registry.into();
-
-        let type_gen = TypeGenerator::new(
-            &portable_types,
-            "root",
-            Default::default(),
-            Default::default(),
-        );
-        let types = type_gen.generate_types_mod();
-        let tests_mod = get_mod(&types, MOD_PATH).unwrap();
-
-        assert_eq!(
-            tests_mod.into_token_stream().to_string(),
-            quote! {
-                pub mod tests {
-                    use super::root;
-                    #[derive(::codec::Encode, ::codec::Decode)]
-                    pub enum E {
-                        A(std::boxed::Box<bool>,),
-                        B { a: std::boxed::Box<u32>, },
-                    }
-                }
-            }
-            .to_string()
-        )
-    }
-
-    #[test]
-    fn range_fields() {
-        #[allow(unused)]
-        #[derive(TypeInfo)]
-        struct S {
-            a: core::ops::Range<u32>,
-            b: core::ops::RangeInclusive<u32>,
-        }
-
-        let mut registry = Registry::new();
-        registry.register_type(&meta_type::<S>());
-        let portable_types: PortableRegistry = registry.into();
-
-        let type_gen = TypeGenerator::new(
-            &portable_types,
-            "root",
-            Default::default(),
-            Default::default(),
-        );
-        let types = type_gen.generate_types_mod();
-        let tests_mod = get_mod(&types, MOD_PATH).unwrap();
-
-        assert_eq!(
-            tests_mod.into_token_stream().to_string(),
-            quote! {
-                pub mod tests {
-                    use super::root;
-                    #[derive(::codec::Encode, ::codec::Decode)]
-                    pub struct S {
-                        pub a: ::core::ops::Range<u32>,
-                        pub b: ::core::ops::RangeInclusive<u32>,
-                    }
-                }
-            }
-            .to_string()
-        )
-    }
-
-    #[test]
-    fn generics() {
-        #[allow(unused)]
-        #[derive(TypeInfo)]
-        struct Foo<T> {
-            a: T,
-        }
-
-        #[allow(unused)]
-        #[derive(TypeInfo)]
-        struct Bar {
-            b: Foo<u32>,
-            c: Foo<u8>,
-        }
-
-        let mut registry = Registry::new();
-        registry.register_type(&meta_type::<Bar>());
-        let portable_types: PortableRegistry = registry.into();
-
-        let type_gen = TypeGenerator::new(
-            &portable_types,
-            "root",
-            Default::default(),
-            Default::default(),
-        );
-        let types = type_gen.generate_types_mod();
-        let tests_mod = get_mod(&types, MOD_PATH).unwrap();
-
-        assert_eq!(
-            tests_mod.into_token_stream().to_string(),
-            quote! {
-                pub mod tests {
-                    use super::root;
-                    #[derive(::codec::Encode, ::codec::Decode)]
-                    pub struct Bar {
-                        pub b: root::subxt_codegen::types::tests::Foo<u32>,
-                        pub c: root::subxt_codegen::types::tests::Foo<u8>,
-                    }
-                    #[derive(::codec::Encode, ::codec::Decode)]
-                    pub struct Foo<_0> {
-                        pub a: _0,
-                    }
-                }
-            }
-            .to_string()
-        )
-    }
-
-    #[test]
-    fn generics_nested() {
-        #[allow(unused)]
-        #[derive(TypeInfo)]
-        struct Foo<T, U> {
-            a: T,
-            b: Option<(T, U)>,
-        }
-
-        #[allow(unused)]
-        #[derive(TypeInfo)]
-        struct Bar<T> {
-            b: Foo<T, u32>,
-        }
-
-        let mut registry = Registry::new();
-        registry.register_type(&meta_type::<Bar<bool>>());
-        let portable_types: PortableRegistry = registry.into();
-
-        let type_gen = TypeGenerator::new(
-            &portable_types,
-            "root",
-            Default::default(),
-            Default::default(),
-        );
-        let types = type_gen.generate_types_mod();
-        let tests_mod = get_mod(&types, MOD_PATH).unwrap();
-
-        assert_eq!(
-            tests_mod.into_token_stream().to_string(),
-            quote! {
-                pub mod tests {
-                    use super::root;
-                    #[derive(::codec::Encode, ::codec::Decode)]
-                    pub struct Bar<_0> {
-                        pub b: root::subxt_codegen::types::tests::Foo<_0, u32>,
-                    }
-
-                    #[derive(::codec::Encode, ::codec::Decode)]
-                    pub struct Foo<_0, _1> {
-                        pub a: _0,
-                        pub b: Option<(_0, _1,)>,
-                    }
-                }
-            }
-            .to_string()
-        )
-    }
-
-    #[test]
-    fn generate_bitvec() {
-        use bitvec::{
-            order::{
-                Lsb0,
-                Msb0,
-            },
-            vec::BitVec,
-        };
-
-        #[allow(unused)]
-        #[derive(TypeInfo)]
-        struct S {
-            lsb: BitVec<Lsb0, u8>,
-            msb: BitVec<Msb0, u16>,
-        }
-
-        let mut registry = Registry::new();
-        registry.register_type(&meta_type::<S>());
-        let portable_types: PortableRegistry = registry.into();
-
-        let type_gen = TypeGenerator::new(
-            &portable_types,
-            "root",
-            Default::default(),
-            Default::default(),
-        );
-        let types = type_gen.generate_types_mod();
-        let tests_mod = get_mod(&types, MOD_PATH).unwrap();
-
-        assert_eq!(
-            tests_mod.into_token_stream().to_string(),
-            quote! {
-                pub mod tests {
-                    use super::root;
-                    #[derive(::codec::Encode, ::codec::Decode)]
-                    pub struct S {
-                        pub lsb: ::subxt::bitvec::vec::BitVec<root::bitvec::order::Lsb0, u8>,
-                        pub msb: ::subxt::bitvec::vec::BitVec<root::bitvec::order::Msb0, u16>,
-                    }
-                }
-            }
-            .to_string()
-        )
-    }
-
-    #[test]
-    fn generics_with_alias_adds_phantom_data_marker() {
-        trait Trait {
-            type Type;
-        }
-
-        impl Trait for bool {
-            type Type = u32;
-        }
-
-        type Foo<T> = <T as Trait>::Type;
-        type Bar<T, U> = (<T as Trait>::Type, <U as Trait>::Type);
-
-        #[allow(unused)]
-        #[derive(TypeInfo)]
-        struct NamedFields<T: Trait> {
-            b: Foo<T>,
-        }
-
-        #[allow(unused)]
-        #[derive(TypeInfo)]
-        struct UnnamedFields<T: Trait, U: Trait>(Bar<T, U>);
-
-        let mut registry = Registry::new();
-        registry.register_type(&meta_type::<NamedFields<bool>>());
-        registry.register_type(&meta_type::<UnnamedFields<bool, bool>>());
-        let portable_types: PortableRegistry = registry.into();
-
-        let type_gen = TypeGenerator::new(
-            &portable_types,
-            "root",
-            Default::default(),
-            Default::default(),
-        );
-        let types = type_gen.generate_types_mod();
-        let tests_mod = get_mod(&types, MOD_PATH).unwrap();
-
-        assert_eq!(
-            tests_mod.into_token_stream().to_string(),
-            quote! {
-                pub mod tests {
-                    use super::root;
-                    #[derive(Debug, Eq, PartialEq, ::codec::Encode, ::codec::Decode, ::codec::CompactAs)]
-                    pub struct NamedFields<_0> {
-                        pub b: u32,
-                        #[codec(skip)] pub __subxt_unused_type_params: ::core::marker::PhantomData<_0>,
-                    }
-                    #[derive(::codec::Encode, ::codec::Decode)]
-                    pub struct UnnamedFields<_0, _1> (
-                        pub (u32, u32,),
-                        #[codec(skip)] pub ::core::marker::PhantomData<(_0, _1)>,
-                    );
-                }
-            }
-            .to_string()
-        )
-    }
-
-    #[test]
-    fn modules() {
-        mod modules {
-            pub mod a {
-                #[allow(unused)]
-                #[derive(scale_info::TypeInfo)]
-                pub struct Foo {}
-
-                pub mod b {
-                    #[allow(unused)]
-                    #[derive(scale_info::TypeInfo)]
-                    pub struct Bar {
-                        a: super::Foo,
-                    }
-                }
-            }
-
-            pub mod c {
-                #[allow(unused)]
-                #[derive(scale_info::TypeInfo)]
-                pub struct Foo {
-                    a: super::a::b::Bar,
-                }
-            }
-        }
-
-        let mut registry = Registry::new();
-        registry.register_type(&meta_type::<modules::c::Foo>());
-        let portable_types: PortableRegistry = registry.into();
-
-        let type_gen = TypeGenerator::new(
-            &portable_types,
-            "root",
-            Default::default(),
-            Default::default(),
-        );
-        let types = type_gen.generate_types_mod();
-        let tests_mod = get_mod(&types, MOD_PATH).unwrap();
-
-        assert_eq!(
-            tests_mod.into_token_stream().to_string(),
-            quote! {
-                pub mod tests {
-                    use super::root;
-                    pub mod modules {
-                        use super::root;
-                        pub mod a {
-                            use super::root;
-
-                            pub mod b {
-                                use super::root;
-
-                                #[derive(::codec::Encode, ::codec::Decode)]
-                                pub struct Bar {
-                                    pub a: root::subxt_codegen::types::tests::modules::a::Foo,
-                                }
-                            }
-
-                            #[derive(::codec::Encode, ::codec::Decode)]
-                            pub struct Foo {}
-                        }
-
-                        pub mod c {
-                            use super::root;
-
-                            #[derive(::codec::Encode, ::codec::Decode)]
-                            pub struct Foo {
-                                pub a: root::subxt_codegen::types::tests::modules::a::b::Bar,
-                            }
-                        }
-                    }
-                }
-            }
-            .to_string()
-        )
     }
 }
